+++ src/gui/aggregatoragri.cpp
@@ -356,6 +356,7 @@ void AggregatorAgri::onUpdateGrove( unsigned int id )
 			m_groveInfo.plantTrees = props.plant;
 			m_groveInfo.fellTrees  = props.fell;
 			m_groveInfo.treeType   = props.treeType;
+			m_groveInfo.planted	   = grove->numTrees();
 
 			if ( m_groveInfo.product.plantID != m_groveInfo.treeType )
 			{